tcgtool: new recipe
Create recipe for tcgtool, a program that replicates the structures used
to represent data measured and hashed to extend TPM PCRs.

This is useful to compute a PCR hash at runtime, which is normally
computed by the firmware before the OS boots. This allows for adjusting
a TPM2 policy to unlock the disk encryption passphrase with the updated
state on the next boot.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>