Move policy update to HUP commit hook
When migrating the TPM2 policy used to secure the LUKS passphrase to use
different PCRs, we temporarily want to maintain fallback capability in
case the newly installed hostapp doesn't pass healthchecks. This allows
the system to boot back into the original OS and try again.

In order to do so, we leave the passphrase in place with the old PCR
authentication policy. The cryptsetup hook in the initramfs will try
PCRs 0,2,3,7 and if those don't work we fallback to the original PCRs.

Once the new system successfully boots, we'll re-encrypt the passphrase
and use the new PCRs to create a policy to secure the key.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>